add notes for more cleanup
