+++ lib/parse.js
@@ -4,6 +4,8 @@ import ipapi from 'ipapi.co'
 
 import computeCredit from '../lib/latepenalty'
 
+// TODO refactor/split this out
+
 export function parseCredit({ dueDate, finishDate = moment() }) {
   const diff = moment(finishDate).diff(dueDate, 'seconds')
   const cred = computeCredit(diff)
@@ -59,31 +61,6 @@ export default function parsePromiseWithIp({ urtext, ip }) {
       reject('failed to parse timezone from IP') // new Error('text')
     })
   })
-}
-
-export function parsePromise({ urtext, timezone }) {  
-  if (!urtext) return;
-  urtext = urtext.replace(/^\/|\/$/g, '') // remove leading/trailing slashes
-  urtext = urtext.replace(/\/edit\/?/, '')
-  const [, user, domain, slug, secret] = urtext.match(/^([^\.]+)\.([^\.\/]+\.[^\/]+)\/?([^\?]*)(.*)$/)
-
-  if (slug.length > 0) {
-    const [, text, , modifier, due] = slug.match(/^(.*?)(\/(by|in|at)\/|$)(.*)$/)
-
-    const promise = {
-      id: urtext.toLowerCase(), // prevent duplicates from capitalization differences
-      urtext: urtext.replace(/\?(.*)=?(.*)$/, ''),
-      what: parseText(text),
-      tdue: parseDate({ modifier, due, timezone }),
-      user,
-      domain,
-      slug,
-      secret: secret.substr(2)
-    }
-    
-    console.log('parsePromise', urtext, promise)
-    return promise
-  }
 }
 
 export function parseText(text) {
